// src/app/models/hero-prop.ts
export interface HeroProp {
  id: number;
  name: string;
  subText: string;
  textWeight: string;
  textStyle: string;
  subTextSize: string;
  subTextWeight: string;
  subTextStyle: string;
  textColor: string;
  backgroundColor: string;
  sizeY: number;
  image: string;
  imageTintColor: string;
  imageDarken: string;
  imageToggle: boolean;
  alignText: string;
  justifyText: string;
}